Scan for networks on every call to activate
This uses wifiscanner (which in turn uses iw) to scan for networks. This
means we can refresh our list of networks while still providing an AP
via NetworkManager.

Unfortunately, wifiscanner does not currently parse security for
networks on Linux, so we can't use its results. Fortunately, just doing
the scan seems to be enough for NetworkManager to pick up the results!